@@ -1,25 +1,37 @@
 import React, { useState } from 'react';
-
 import Button from '../../UI/Button/Button';
 import './CourseInput.css';
 
-const CourseInput = props => {
+const CourseInput = (props) => {
   const [enteredValue, setEnteredValue] = useState('');
+  const [isValid, setIsValid] = useState(true);
 
-  const goalInputChangeHandler = event => {
+  const goalInputChangeHandler = (event) => {
     setEnteredValue(event.target.value);
   };
 
-  const formSubmitHandler = event => {
+  const formSubmitHandler = (event) => {
     event.preventDefault();
+    if (enteredValue.trim().length === 0) {
+      setIsValid(false);
+      return;
+    }
     props.onAddGoal(enteredValue);
+    setIsValid(true);
   };
 
   return (
     <form onSubmit={formSubmitHandler}>
       <div className="form-control">
-        <label>Course Goal</label>
-        <input type="text" onChange={goalInputChangeHandler} />
+        <label style={{ color: !isValid ? 'red' : 'black' }}>Course Goal</label>
+        <input
+          style={{
+            borderColor: !isValid ? 'red' : 'black',
+            backgroundColor: isValid ? 'transparent' : 'salmon',
+          }}
+          type="text"
+          onChange={goalInputChangeHandler}
+        />
       </div>
       <Button type="submit">Add Goal</Button>
     </form>
